E2E testing
The e2e test is working as expected with a local container registry as the destination registry.
TLS_VERIFY parameter added in the task-sc yaml file which is to be set as false while using local container registry.
The source and destination image digests are stored in results and finally compared using regex matching.

Signed-off-by: Senjuti De <sde@redhat.com>